Add ColorInputPanel test
